Variable support

+++ NitroComposer/SequenceCommands/Rand/ADSRRandCommand.cs
@@ -1,9 +1,9 @@
 ﻿namespace Nitro.Composer.SequenceCommands.Rand {
 	public class ADSRRandCommand : BaseSequenceCommand {
-		public uint Min, Max;
+		public short Min, Max;
 		public ADSRCommand.EnvPos envPos;
 
-		public ADSRRandCommand(ADSRCommand.EnvPos envPos, uint min, uint max) {
+		public ADSRRandCommand(ADSRCommand.EnvPos envPos, short min, short max) {
 			this.envPos = envPos;
 			Min = min;
 			Max = max;
