+++ NitroComposer/SequenceCommands/Rand/ModulationRandCommand.cs
@@ -1,9 +1,9 @@
 ﻿namespace Nitro.Composer.SequenceCommands.Rand {
 	public class ModulationRandCommand : BaseSequenceCommand {
-		public uint Min, Max;
+		public short Min, Max;
 		public ModulationCommand.ModType Type;
 
-		public ModulationRandCommand(ModulationCommand.ModType type, uint min, uint max) {
+		public ModulationRandCommand(ModulationCommand.ModType type, short min, short max) {
 			Type = type;
 			Min = min;
 			Max = max;
