+++ NitroComposer/SequenceCommands/Rand/PitchBendRandCommand.cs
@@ -1,9 +1,9 @@
 ﻿namespace Nitro.Composer.SequenceCommands.Rand {
 	public class PitchBendRandCommand : BaseSequenceCommand {
-		public uint BendMin, BendMax;
+		public short BendMin, BendMax;
 		public bool IsRange;
 
-		public PitchBendRandCommand(uint bendMin, uint bendMax, bool range) {
+		public PitchBendRandCommand(short bendMin, short bendMax, bool range) {
 			BendMin = bendMin;
 			BendMax = bendMax;
 			IsRange = range;
