+++ NitroComposer/SequenceCommands/Rand/PortamentoTimeRandCommand.cs
@@ -1,7 +1,7 @@
 ﻿namespace Nitro.Composer.SequenceCommands.Rand {
 	public class PortamentoTimeRandCommand : BaseSequenceCommand {
-		public uint TimeMin, TimeMax;
-		public PortamentoTimeRandCommand(uint min, uint max) {
+		public short TimeMin, TimeMax;
+		public PortamentoTimeRandCommand(short min, short max) {
 			TimeMin = min;
 			TimeMax = max;
 		}
