+++ NitroComposer/SequenceCommands/Rand/VarRandCommand.cs
@@ -2,9 +2,9 @@
 	public class VarRandCommand : BaseSequenceCommand {
 		public VarCommand.Operator Op;
 		public byte Variable;
-		public uint OperandMin, OperandMax;
+		public short OperandMin, OperandMax;
 
-		public VarRandCommand(VarCommand.Operator Op, byte variable, uint min, uint max) {
+		public VarRandCommand(VarCommand.Operator Op, byte variable, short min, short max) {
 			this.Op = Op;
 			this.Variable = variable;
 			this.OperandMin = min;
