+++ NitroComposer/SequenceCommands/Rand/VolumeRandCommand.cs
@@ -1,9 +1,9 @@
 ﻿namespace Nitro.Composer.SequenceCommands.Rand {
 	public class VolumeRandCommand : BaseSequenceCommand {
 		public bool Master;
-		public uint VolumeMin, VolumeMax;
+		public short VolumeMin, VolumeMax;
 
-		public VolumeRandCommand(uint min, uint max, bool master) {
+		public VolumeRandCommand(short min, short max, bool master) {
 			VolumeMin = min;
 			VolumeMax = max;
 			Master = master;
